refactor transifex/pull script
LEARNER-1350 (subtask of LEARNER-1089)
