--- transifex/settings.yaml
@@ -1,5 +1,0 @@
-repos:
-  - git@github.com:edx/course-discovery.git
-  - git@github.com:edx/credentials.git
-  - git@github.com:edx/ecommerce.git
-  - git@github.com:edx/programs.git
